Refactor variant scoring and person ID sorting
