+++ bin/anthology/people.py
@@ -15,7 +15,28 @@
 # limitations under the License.
 
 import logging as log
+from functools import cached_property
+import re
+from slugify import slugify
 import anthology.formatter as my_formatter
+
+
+def score_variant(name):
+    """Heuristically assign scores to names, with the idea of assigning higher
+    scores to spellings more likely to be the correct canonical variant."""
+    name = repr(name)
+    # Prefer longer variants
+    score = len(name)
+    # Prefer variants with non-ASCII characters
+    score += sum((ord(c) > 127) for c in name)
+    # Penalize upper-case characters after word boundaries
+    score -= sum(any(c.isupper() for c in w[1:]) for w in re.split(r"\W+", name))
+    # Penalize lower-case characters at word boundaries
+    score -= sum(w[0].islower() if w else 0 for w in re.split(r"\W+", name))
+    if name[0].islower():  # extra penalty for first name
+        score -= 1
+
+    return score
 
 
 class PersonName:
@@ -65,7 +86,7 @@ class PersonName:
         last = dict_["last"]
         return PersonName(first, last)
 
-    @property
+    @cached_property
     def full(self):
         """
         Return the full rendering of the name.
@@ -81,6 +102,22 @@ class PersonName:
             return f"{form} ({self.variant.full})"
         else:
             return form
+
+    @cached_property
+    def slug(self):
+        # This is effectively used as the person's "id".
+        # If two names slugify to the same thing, we assume they are the same person.
+        # This happens when there are missing accents in one version, or
+        # when we have an inconsistent first/last split for multiword names.
+        # These cases have in practice always referred to the same person.
+        slug = slugify(repr(self))
+        if not slug:
+            slug = "none"
+        return slug
+
+    @cached_property
+    def score(self):
+        return score_variant(self)
 
     @property
     def id_(self):
